+++ ipEngine/returnControlStackFrame.go
@@ -7,26 +7,26 @@ package ipEngine
 import "khalehla/pkg"
 
 type ReturnControlStackFrame struct {
-	bankLevel             uint
-	bankDescriptorIndex   uint
-	offset                uint
+	bankLevel             uint64
+	bankDescriptorIndex   uint64
+	offset                uint64
 	trapFlag              bool
-	basicModeBaseRegister uint                    // For return to basic mode, add 12 to this to find the base register for return
+	basicModeBaseRegister uint64                  // For return to basic mode, add 12 to this to find the base register for return
 	designatorRegister    *pkg.DesignatorRegister //	only bits 12-17 are significant
 	accessKey             *pkg.AccessKey
 }
 
-func (rcsf *ReturnControlStackFrame) SetBankLevel(value uint) *ReturnControlStackFrame {
+func (rcsf *ReturnControlStackFrame) SetBankLevel(value uint64) *ReturnControlStackFrame {
 	rcsf.bankLevel = value & 07
 	return rcsf
 }
 
-func (rcsf *ReturnControlStackFrame) SetBankDescriptorIndex(value uint) *ReturnControlStackFrame {
+func (rcsf *ReturnControlStackFrame) SetBankDescriptorIndex(value uint64) *ReturnControlStackFrame {
 	rcsf.bankDescriptorIndex = value & 077777
 	return rcsf
 }
 
-func (rcsf *ReturnControlStackFrame) SetOffset(value uint) *ReturnControlStackFrame {
+func (rcsf *ReturnControlStackFrame) SetOffset(value uint64) *ReturnControlStackFrame {
 	rcsf.offset = value & 0_777777
 	return rcsf
 }
@@ -36,7 +36,7 @@ func (rcsf *ReturnControlStackFrame) SetTrapFlag(value bool) *ReturnControlStack
 	return rcsf
 }
 
-func (rcsf *ReturnControlStackFrame) SetBasicModeBaseRegister(value uint) *ReturnControlStackFrame {
+func (rcsf *ReturnControlStackFrame) SetBasicModeBaseRegister(value uint64) *ReturnControlStackFrame {
 	rcsf.basicModeBaseRegister = value & 03
 	return rcsf
 }
@@ -68,11 +68,11 @@ func (rcsf *ReturnControlStackFrame) WriteToBuffer(buffer []pkg.Word36) {
 }
 
 func NewReturnControlStackFrameFromComponents(
-	bankLevel uint,
-	bankDescriptorIndex uint,
-	offset uint,
+	bankLevel uint64,
+	bankDescriptorIndex uint64,
+	offset uint64,
 	trapFlag bool,
-	basicModeBaseRegister uint,
+	basicModeBaseRegister uint64,
 	designatorRegister *pkg.DesignatorRegister,
 	accessKey *pkg.AccessKey) *ReturnControlStackFrame {
 	rcsf := ReturnControlStackFrame{}
@@ -88,12 +88,12 @@ func NewReturnControlStackFrameFromComponents(
 
 func NewReturnControlStackFrameFromBuffer(source []pkg.Word36) *ReturnControlStackFrame {
 	return &ReturnControlStackFrame{
-		bankLevel:             uint(source[0] >> 33),
-		bankDescriptorIndex:   uint(source[0]>>18) & 077777,
-		offset:                uint(source[0] & 0777777),
+		bankLevel:             uint64(source[0] >> 33),
+		bankDescriptorIndex:   uint64(source[0]>>18) & 077777,
+		offset:                uint64(source[0] & 0777777),
 		trapFlag:              (source[1] >> 35) == 01,
-		basicModeBaseRegister: uint(source[1]>>24) & 03,
+		basicModeBaseRegister: uint64(source[1]>>24) & 03,
 		designatorRegister:    pkg.NewDesignatorRegisterFromComposite(uint64(source[1] & 0_000077_000000)),
-		accessKey:             pkg.NewAccessKeyFromComposite(uint(source[1])),
+		accessKey:             pkg.NewAccessKeyFromComposite(uint64(source[1])),
 	}
 }
