+++ tasm/bank.go
@@ -11,7 +11,7 @@ import (
 
 type Bank struct {
 	bankDescriptor      *pkg.BankDescriptor
-	bankDescriptorIndex uint
+	bankDescriptorIndex uint64
 	code                []uint64 //	key is L,BDI (18 bits), value is the table of binary 36-bit values to comprise the bank
 }
 
@@ -19,7 +19,7 @@ func (b *Bank) GetBankDescriptor() *pkg.BankDescriptor {
 	return b.bankDescriptor
 }
 
-func (b *Bank) GetBankDescriptorIndex() uint {
+func (b *Bank) GetBankDescriptorIndex() uint64 {
 	return b.bankDescriptorIndex
 }
 
@@ -27,11 +27,11 @@ func (b *Bank) GetCode() []uint64 {
 	return b.code
 }
 
-func (b *Bank) GetCodeLength() uint {
-	return uint(len(b.code))
+func (b *Bank) GetCodeLength() uint64 {
+	return uint64(len(b.code))
 }
 
-func NewBank(bd *pkg.BankDescriptor, bdi uint, code []uint64) *Bank {
+func NewBank(bd *pkg.BankDescriptor, bdi uint64, code []uint64) *Bank {
 	return &Bank{
 		bankDescriptor:      bd,
 		bankDescriptorIndex: bdi,
